+++ chemreac/__init__.py
@@ -12,7 +12,4 @@ from __future__ import absolute_import, division, print_function
 
 from ._release import __version__
 
-from .core import (
-    ReactionDiffusion, DENSE, BANDED, SPARSE, FLAT,
-    CYLINDRICAL, SPHERICAL, Geom_names
-)
+from .core import ReactionDiffusion, Geom_names