+++ examples/steady_state.py
@@ -8,9 +8,7 @@ from math import log
 import argh
 import numpy as np
 
-from chemreac import (
-    ReactionDiffusion, FLAT, CYLINDRICAL, SPHERICAL
-)
+from chemreac import ReactionDiffusion
 from chemreac.integrate import run
 from chemreac.util.plotting import plot_solver_linear_error
 
@@ -56,7 +54,6 @@ def integrate_rd(D=2e-3, t0=3., tend=7., x0=0.0, xend=1.0, mu=None, N=32,
     tout = np.linspace(t0, tend, nt)
 
     assert geom in 'fcs'
-    geom = {'f': FLAT, 'c': CYLINDRICAL, 's': SPHERICAL}[geom]
 
     # Setup the grid
     _x0 = log(x0) if logx else x0
@@ -68,9 +65,9 @@ def integrate_rd(D=2e-3, t0=3., tend=7., x0=0.0, xend=1.0, mu=None, N=32,
     mob = 0.3
     # Initial conditions
     y0 = {
-        FLAT: y0_flat_cb,
-        CYLINDRICAL: y0_cylindrical_cb,
-        SPHERICAL: y0_spherical_cb
+        'f': y0_flat_cb,
+        'c': y0_cylindrical_cb,
+        's': y0_spherical_cb
     }[geom](x, logx)
 
     # Setup the system
@@ -80,34 +77,6 @@ def integrate_rd(D=2e-3, t0=3., tend=7., x0=0.0, xend=1.0, mu=None, N=32,
 
     assert not lrefl
     assert not rrefl
-    # bin_k_factor = [[] for _ in range(N)]
-    # bin_k_factor_span = []
-    # if lrefl:
-    #     # isolated sys is not stationary for linear conc profile with finite
-    #     # slope, hence we add production reaction at left boundary
-    #     assert nstencil == 3  # k is derived for parabola through -x0, x0, x1
-    #     stoich_active.append([0])
-    #     stoich_prod.append([0, 0])
-    #     for i in range(N):
-    #         bin_k_factor[i].append(1 if i == 0 else 0)
-    #     bin_k_factor_span.append(1)
-    #     x0 = (x[0]+x[1])/2 - x[0]
-    #     x1 = (x[1]+x[2])/2 - x[0]
-    #     C = {FLAT: 2, CYLINDRICAL: 4, SPHERICAL: 6}[geom]
-    #     k.append(C*D*(y0[0]-y0[1])/(y0[0]*(x1**2 - x0**2)))
-    # if rrefl:
-    #     # for same reason as above, a consumption reaction is added at right
-    #     # boundary
-    #     assert nstencil == 3
-    #     stoich_active.append([0])
-    #     stoich_prod.append([])
-    #     for i in range(N):
-    #         bin_k_factor[i].append(1 if i == (N-1) else 0)
-    #     bin_k_factor_span.append(1)
-    #     xNm1 = (x[N-1]+x[N])/2-x[N]
-    #     xNm2 = (x[N-2]+x[N-1])/2-x[N]
-    #     C = {FLAT: 2, CYLINDRICAL: 4, SPHERICAL: 6}[geom]
-    #     k.append(C*D*(y0[N-2]-y0[N-1])/(y0[N-1]*(xNm2**2 - xNm1**2)))
 
     rd = ReactionDiffusion(
         n, stoich_active, stoich_prod, k, N,
@@ -115,8 +84,6 @@ def integrate_rd(D=2e-3, t0=3., tend=7., x0=0.0, xend=1.0, mu=None, N=32,
         z_chg=[1],
         mobility=[mob],
         x=x,
-        # bin_k_factor=bin_k_factor,
-        # bin_k_factor_span=bin_k_factor_span,
         geom=geom,
         logy=logy,
         logt=logt,
@@ -127,7 +94,7 @@ def integrate_rd(D=2e-3, t0=3., tend=7., x0=0.0, xend=1.0, mu=None, N=32,
     )
 
     if efield:
-        if geom != FLAT:
+        if geom != 'f':
             raise ValueError("Only analytic sol. for flat drift implemented.")
         rd.efield = efield_cb(rd.xcenters, logx)